Add files via upload
The Hill cipher algorithm is a symmetric key cryptographic technique used for encryption and decryption of data. It was invented by Lester S. Hill in 1929 and is based on linear algebra.

In this algorithm, plaintext is divided into blocks of fixed size, which are then converted into numeric values according to a predetermined mapping. These numeric values are then multiplied by a matrix, which is the key of the cipher, to produce the ciphertext.

To decrypt the ciphertext, the inverse of the key matrix is multiplied with the ciphertext matrix to get back the original plaintext values. These numeric values are then converted back to their original format to retrieve the original plaintext.

The security of the Hill cipher algorithm depends on the size of the key matrix, and it can be vulnerable to certain attacks, such as known plaintext attacks and chosen plaintext attacks. Therefore, it is typically used for educational purposes rather than for secure communications.